% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Country_Search.R
\name{Country_Search}
\alias{Country_Search}
\title{Country Search}
\usage{
Country_Search(country = df$CountryCode)
}
\arguments{
\item{country}{This should be the 2-letter country code for the country of interest.
If no country is specified, the function returns a data frame of the information for all countries.}
}
\value{
This function returns the new and total confirmed cases, new and total deaths, and new and total recoveries for the country inputted by the end user.
}
\description{
Country Search
}
\examples{
Country_Search("US")
##to find Covid-19 information for the rate in the United States.

}
