% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Death_Rate.R
\name{Death_Rate}
\alias{Death_Rate}
\title{Death Rate}
\usage{
Death_Rate(country = df$CountryCode)
}
\arguments{
\item{country}{This should be the 2-letter country code for the country of interest.
If no country is specified, the function returns a data frame of the death rates for all countries.}
}
\value{
This function returns the Covid-19 death rate (percentage) for the country inputted by the end user.
}
\description{
Death Rate
}
\examples{
Death_Rate("US")
##to find the death rate in the United States.

}
