% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New_Deaths_Country_Stats.R
\name{New_Deaths_Country_Stats}
\alias{New_Deaths_Country_Stats}
\title{New Deaths Country Statistics}
\usage{
New_Deaths_Country_Stats(stat)
}
\arguments{
\item{stat}{This should be the statistic the end user wants to analyze.
It can be "maximum", "minimum", "average" or "sum".}
}
\value{
The output depends on the stat parameter input.
If the stat is the maximum or minimum, the function outputs a data frame of the name and number of new Covid-19 deaths for the country/countries with the most or fewest new deaths, respectively.
If the stat is the average, it outputs the average number of new Covid-19 deaths by a world country.
If the stat is the sum, it outputs the sum of all the world countries' new Covid-19 deaths.
}
\description{
New Deaths Country Statistics
}
\examples{
New_Deaths_Country_Stats("minimum")
##to find the country with the fewest new Covid-19 deaths.

}
