% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recovery_Rate.R
\name{Recovery_Rate}
\alias{Recovery_Rate}
\title{Recovery Rate}
\usage{
Recovery_Rate(country = df$CountryCode)
}
\arguments{
\item{country}{This should be the 2-letter country code for the country of interest.
If no country is specified, the function returns a data frame of the recovery rates for all countries.}
}
\value{
This function returns the Covid-19 recovery rate (percentage) for the country inputted by the end user.
}
\description{
Recovery Rate
}
\examples{
Recovery_Rate("US")
##to find the recovery rate in the United States.

}
