% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Deaths_Country_Stats.R
\name{Total_Deaths_Country_Stats}
\alias{Total_Deaths_Country_Stats}
\title{Total Deaths Country Statistics}
\usage{
Total_Deaths_Country_Stats(stat)
}
\arguments{
\item{stat}{This should be the statistic the end user wants to analyze.
It can be "maximum", "minimum", "average" or "sum".}
}
\value{
The output depends on the stat parameter input.
If the stat is the maximum or minimum, the function outputs a data frame of the name and total Covid-19 deaths for the country/countries with the most or fewest total deaths, respectively.
If the stat is the average, it outputs the average total Covid-19 deaths by a world country.
If the stat is the sum, it outputs the sum of all the world countries' total Covid-19 deaths.
}
\description{
Total Deaths Country Statistics
}
\examples{
Total_Deaths_Country_Stats("maximum")
##to find the country with the most total Covid-19 deaths.

}
