% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Recoveries_Country_Stats.R
\name{Total_Recoveries_Country_Stats}
\alias{Total_Recoveries_Country_Stats}
\title{Total Recoveries Country Statistics}
\usage{
Total_Recoveries_Country_Stats(stat)
}
\arguments{
\item{stat}{This should be the statistic the end user wants to analyze.
It can be "maximum", "minimum", "average" or "sum".}
}
\value{
The output depends on the stat parameter input.
If the stat is the maximum or minimum, the function outputs a data frame of the name and total Covid-19 recoveries for the country/countries with the most or fewest total recoveries, respectively.
If the stat is the average, it outputs the average total Covid-19 recoveries by a world country.
If the stat is the sum, it outputs the sum of all the world countries' total Covid-19 recoveries.
}
\description{
Total Recoveries Country Statistics
}
\examples{
Total_Recoveries_Country_Stats("sum")
##to find the total number of Covid-19 recoveries in the world.

}
